Fix integer overflow

+++ include/memkind/internal/tachanka.h
@@ -54,8 +54,8 @@ struct ttype {
     __u64 t1;   // start of current window
     __u64 t0;   // timestamp of last processed data
 
-    int n2;   // num of access in prev window
-    int n1;   // num of access in current window
+    double n2;   // num of access in prev window
+    double n1;   // num of access in current window
 
     tachanka_touch_callback touchCb;
     void *touchCbArg;
